requirements.txt
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
linkpreview==0.8.3
neonize==0.0.7
phonenumbers==8.13.29
pillow==10.2.0
protobuf==4.25.2
python-dotenv==1.0.1
python-magic==0.4.27
requests==2.31.0
segno==1.6.0
soupsieve==2.5
tqdm==4.66.1
urllib3==2.2.0
watchdog==3.0.0
